Inference

Inference

Inference

Point estimation

Maximum likelihood estimator

Unbiased estimator

Optimal compatible estimator

P values

Confidence intervals

Naive Clopper-Pearson confidence interval

Clopper-Pearson confidence interval

Minimum mean width confidence interval

Point estimation

BinaryTwoStageDesigns.BinaryTwoStageDesignEstimator — Type.

BinaryTwoStageDesignEstimator

Abstract base type for all estimators.

source

BinaryTwoStageDesigns.estimate — Method.

estimate{T<:Integer}(estimator::BinaryTwoStageDesignEstimator, x1::T, x2::T)

Estimate the response rate from observed x1 and x2.

Parameters

Parameter	Description

estimator	any BinaryTwoStageDeisgnEstimator object

x1	stage-one responses

x2	stage-two responses

Return Value

Real, estimated response rate.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> estimate(est, 0, 0)

source

BinaryTwoStageDesigns.bias — Method.

bias{T<:Real}(estimator::BinaryTwoStageDesignEstimator, p::T)

Bias of estimator given response rate p.

Parameters

Parameter	Description

estimator	any BinaryTwoStageDeisgnEstimator object

p0	upper boundary of null hypothesis

Return Value

Real, bias given p.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> bias(est, .3)

source

BinaryTwoStageDesigns.rmse — Method.

rmse{T<:Real}(estimator::BinaryTwoStageDesignEstimator, p::T)

Root mean squared error of estimator given response rate p.

Parameters

Parameter	Description

estimator	any BinaryTwoStageDeisgnEstimator object

p0	upper boundary of null hypothesis

Return Value

Real, RMSE given p.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> rmse(est, .3)

source

Maximum likelihood estimator

BinaryTwoStageDesigns.MaximumLikelihoodEstimator — Type.

MaximumLikelihoodEstimator

MaximumLikelihoodEstimator(design::BinaryTwoStageDesign)

Simple maximum likelihood estimator for response rate p.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MaximumLikelihoodEstimator(design)

source

Unbiased estimator

BinaryTwoStageDesigns.RaoBlackwellizedEstimator — Type.

RaoBlackwellizedEstimator <: BinaryTwoStageDesignEstimator

RaoBlackwellizedEstimator(design::BinaryTwoStageDesign)

Unbiased estimator for response rate p see also:

Kunzmann K, Kieser M. Point estimation and p‐values in phase II adaptive two‐stage designs with a binary endpoint. Statistics in medicine. 2017 Mar 15;36(6):971-84.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = RaoBlackwellizedEstimator(design)

source

Optimal compatible estimator

BinaryTwoStageDesigns.CompatibleEstimator — Type.

CompatibleEstimator <: BinaryTwoStageDesignEstimator

CompatibleEstimator{TS<:MathProgBase.AbstractMathProgSolver}(
    design::BinaryTwoStageDesign,
    solver::TS;
    prior::Function = jeffreysprior(design),
    k = 100
)

Compatible estimator minimizing expected MSE for response rate p see also:

Kunzmann K, Kieser M. Point estimation and p‐values in phase II adaptive two‐stage designs with a binary endpoint. Statistics in medicine. 2017 Mar 15;36(6):971-84.

Parameters

Parameter	Description

design	BinaryTwoStageDesign

solver	MathProgBase solver, must support quadratic expressions

prior	weight function for MSE values at different p, must be of form f(p::Real)::Real

k	number of equally spaced grid-points for evaluation of MSE and prior

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, Gurobi.GurobiSolver())
julia> est = CompatibleEstimator(design, Gurobi.GurobiSolver())

source

P values

BinaryTwoStageDesigns.p — Method.

p{T1<:Integer, T2<:Real}(estimator::BinaryTwoStageDesignEstimator, x1::T1, x2::T1, p0::T2)

Compute the p value after observing (x1, x2) for null hypothesis H0: p <= p0 with respect to ordering induced by estimator.

Parameters

Parameter	Description

estimator	any BinaryTwoStageDeisgnEstimator object

x1	stage-one responses

x2	stage-two responses

p0	upper boundary of null hypothesis

Return Value

Real, p value.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> p(est, 0, 0, .2)

source

BinaryTwoStageDesigns.incompatibleoutcomes — Method.

incompatibleoutcomes(estimator::BinaryTwoStageDesignEstimator)

Find outcomes where the induced p value implies different decision than the underlying design

Parameters

Parameter	Description

estimator	any BinaryTwoStageDeisgnEstimator object

Return Value

Array with respective outcomes

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> incompatibleoutcomes(est)

source

Confidence intervals

BinaryTwoStageDesigns.ConfidenceInterval — Type.

ConfidenceInterval

Abstract base type for all confidence interval types.

source

BinaryTwoStageDesigns.limits — Function.

limits{T<:Integer}(ci::ConfidenceInterval, x1::T, x2::T)

Return the confidence interval limits for observed x1 and x2.

Parameters

Parameter	Description

ci	any ConfidenceInterval object

x1	stage-one responses

x2	stage-two responses

Return Value

Two element Real vector of limits [lower, upper].

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> ci = ClopperPearsonConfidenceInterval(est, confidence = .9)
julia> limits(ci, 0, 0)

source

BinaryTwoStageDesigns.coverage — Function.

coverage{T<:Real}(ci::ConfidenceInterval, p::T; orientation = "overall")

Return coverage of given confidence interval and response rate p.

Parameters

Parameter	Description

ci	any ConfidenceInterval object

p	response rate

orientation	string indicating the coverage type - "overall", "lower", or "upper"

Return Value

Coverage probability given p.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> ci = ClopperPearsonConfidenceInterval(est, confidence = .9)
julia> coverage(ci, .5)

source

BinaryTwoStageDesigns.meanwidth — Function.

meanwidth{T<:Real}(ci::ConfidenceInterval, p::T)

Mean width of given confidence interval and response rate p.

Parameters

Parameter	Description

ci	any ConfidenceInterval object

p	response rate

Return Value

Mean width given p.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> ci = ClopperPearsonConfidenceInterval(est, confidence = .9)
julia> meanwidth(ci, .5)

source

BinaryTwoStageDesigns.meaninterval — Function.

meaninterval{T<:Real}(ci::ConfidenceInterval, p::T)

Mean interval (average limits) of given confidence interval and response rate p.

Parameters

Parameter	Description

ci	any ConfidenceInterval object

p	response rate

Return Value

Mean interval given p.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> ci = ClopperPearsonConfidenceInterval(est, confidence = .9)
julia> meaninterval(ci, .5)

source

BinaryTwoStageDesigns.findinconsistencies — Function.

findinconsistencies{T<:Real}(ci::ConfidenceInterval)

Return outcomes where the confidence interval contradicts the designs test decision.

Parameters

Parameter	Description

ci	any ConfidenceInterval object

p	response rate

Return Value

Mean interval given p.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, solver = Gurobi.GurobiSolver())
julia> est = MLE(design)
julia> ci = ClopperPearsonConfidenceInterval(est, confidence = .9)
julia> findinconsistencies(ci)

source

Naive Clopper-Pearson confidence interval

BinaryTwoStageDesigns.NaiveClopperPearsonConfidenceInterval — Type.

NaiveClopperPearsonConfidenceInterval <: ConfidenceInterval

NaiveClopperPearsonConfidenceInterval{T<:Real}(
    design::BinaryTwoStageDesign;
    confidence::T = .9
)

Naive Clopper-Pearson confidence interval using default ordering.

Parameters

Parameter	Description

estimator	estimator object defining the sample space ordering

confidence	confidence level of the interval

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, Gurobi.GurobiSolver())
julia> est = MaximumLikelihoodEstimator(design, Gurobi.GurobiSolver())
julia> ci = NaiveClopperPearsonConfidenceInterval(est, confidence = .9)

source

Clopper-Pearson confidence interval

BinaryTwoStageDesigns.ClopperPearsonConfidenceInterval — Type.

ClopperPearsonConfidenceInterval <: ConfidenceInterval

ClopperPearsonConfidenceInterval{T<:Real}(
    estimator::BinaryTwoStageDesignEstimator;
    confidence::T = .9,
    k::Integer = 1001
)

Exact Clopper-Pearson type confidence interval based on ordering induced by estimator.

Parameters

Parameter	Description

estimator	estimator object defining the sample space ordering

confidence	confidence level of the interval

k	number of equally spaced grid-points for invertign the test

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, Gurobi.GurobiSolver())
julia> est = MaximumLikelihoodEstimator(design, Gurobi.GurobiSolver())
julia> ci = ClopperPearsonConfidenceInterval(est, confidence = .9)

source

Minimum mean width confidence interval

BinaryTwoStageDesigns.MinimumMeanWidthConfidenceInterval — Type.

MinimumMeanWidthConfidenceInterval <: ConfidenceInterval

MinimumMeanWidthConfidenceInterval{TS<:MathProgBase.AbstractMathProgSolver}(
    estimator::BinaryTwoStageDesignEstimator,
    rho0::Float64,
    prior::Function,
    solver::TS;
    confidence::Float64 = .9,
    ngrid::Int64 = 100
)

Exact confidence interval based on ordering induced by estimator minimizing the expected squared width with respect to weight function prior(p::Real).

Parameters

Parameter	Description

estimator	estimator object defining the sample space ordering

rho0	upper boundary of null hypothesis

confidence	confidence level of the interval

solver	MathProgBase solver used for optimization, must support quadratic expressions

ngrid	number of equally spaced grid-points on which to check coverage

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> interimsamplesize(ss)
julia> design = getoptimaldesign(15, params, Gurobi.GurobiSolver())
julia> est = MaximumLikelihoodEstimator(design, Gurobi.GurobiSolver())
julia> ci = ClopperPearsonConfidenceInterval(est, confidence = .9)

source