Sample Spaces

The Sample Space

Sample space objects can be used to encode the feasible space of binary two-stage designs for optimization.

BinaryTwoStageDesigns.SampleSpace — Type.

SampleSpace

Defines restrictions on the sample space of a two-stage design to be used during optimization.

source

BinaryTwoStageDesigns.interimsamplesizerange — Method.

interimsamplesizerange(ss::SampleSpace)

Return integer vector of valid $n_1$ values.

source

BinaryTwoStageDesigns.maxsamplesize — Method.

maxsamplesize(ss::SampleSpace)

Return maximal overal sample size as integer.

source

BinaryTwoStageDesigns.maxsamplesize — Method.

maxsamplesize(
  ss::SampleSpace{TI,TR}, n1::TI2
) where {TI<:Integer,TR<:Real,TI2<:Integer}

Return maximal sample size given n1 as integer.

source

BinaryTwoStageDesigns.isgroupsequential — Method.

isgroupsequential(ss::SampleSpace)

Indicate whether the sample space is restricted to group-sequential designs.

source