Two-Stage Designs

Binary Two-Stage Designs

BinaryTwoStageDesigns.Design — Type.

Design{T1<:Integer, T2<:Real, PType<:Parameters}

Binary two-stage design object defining a discrete sampe size function $n(x_1)$ and $c(x_1)$, where $x_1$ is the observed number of stage-one events and the test rejects the superiority null hypothesis whenever $x_1 + x_2 > c(x_1)$.

source

BinaryTwoStageDesigns.Design — Method.

Design(n, c, params::PType) where {PType<:Parameters}

Construct a Design object from n and c directly. Here, n must be convertable to an integer vector and c to a real vector (both of same length $n_1$). Then, n[x1 + 1] is the final sample size after observing x1 responses in stage one and c[x1 + 1] the corresponding critical value. A constructor without explicitly specifying a set of parameters with respect to which thee design is optimal also exists.

Parameters

Parameter	Description

n	sample size function (must be convertable to integer vector)

c	critical value function (must be convertable to real vector)

params	parameter object

source

BinaryTwoStageDesigns.Design — Method.

Design(n, c)

Construct a Design object from n and c directly. Here, n must be convertable to an integer vector and c to a real vector (both of same length $n_1$).

Parameters

Parameter	Description

n	sample size function (must be convertable to integer vector)

c	critical value function (must be convertable to real vector)

source

Base.convert — Method.

convert(::Type{DataFrames.DataFrame}, design::Design)

Convert a design to a DataFrame with column names x1, n, and c.

source

BinaryTwoStageDesigns.interimsamplesize — Method.

interimsamplesize(design::Design)

Return interim sample size $n_1$ of a design.

source

BinaryTwoStageDesigns.parameters — Method.

parameters(design::Design)

Return stored parameter object of a design.

source

BinaryTwoStageDesigns.samplesize — Method.

samplesize(design::Design)

Return vector of final sample sizes.

source

BinaryTwoStageDesigns.samplesize — Method.

samplesize(design::Design, x1::T) where {T<:Integer}

Return final sample size for x1.

source

BinaryTwoStageDesigns.criticalvalue — Method.

criticalvalue(design::Design)

Return vector of critical values.

source

BinaryTwoStageDesigns.criticalvalue — Method.

criticalvalue(design::Design, x1::T) where {T<:Integer}

Return critical value for x1.

source

BinaryTwoStageDesigns.power — Method.

power(design::Design, p::T) where {T<:Real}

Power of a design for a given response rate $p$ .

Parameters

Parameter	Description

design	a Design

p	response probability

source

BinaryTwoStageDesigns.power — Method.

power(design::Design, x1::T1, p::T2) where {T1<:Integer, T2<:Real}

conditional power of a design for a given response rate $p$ and the observed number of stage-one responses x1.

Parameters

Parameter	Description

design	a Design

p	response probability

x1	number of stage-one responses

source

BinaryTwoStageDesigns.expectedpower — Method.

expectedpower(
  design::Design, x1::T, prior::Function; mcrv::Real = mcrv(parameters(design))
) where {T<:Integer}

Compute the conditional expected power of a design given x1.

Parameters

Parameter	Description

design	a Design

x1	stage one number of responses

prior	prior function prior(p) for response probability p

mcrv	minimal clinically relevant value for expected power calculation

source

BinaryTwoStageDesigns.expectedpower — Method.

expectedpower(design::Design, prior::Function; mcrv::Real = mcrv(parameters(design)))

Compute the expected power of a given design.

Parameters

Parameter	Description

design	a Design

prior	prior function prior(p) for response probability p

source

BinaryTwoStageDesigns.stoppingforfutility — Method.

stoppingforfutility{T<:Real}(design::Design, p::T) where {T<:Real}

Compute probability of stopping early for futility of a given design.

Parameters

Parameter	Description

design	a Design

p	response probability

source

BinaryTwoStageDesigns.stoppingforefficacy — Method.

stoppingforefficacy{T<:Real}(design::Design, p::T) where {T<:Real}

Compute probability of stopping early for futility of a given design.

Parameters

Parameter	Description

design	a Design

p	response probability

source

BinaryTwoStageDesigns.score — Method.

score(design::Design, params::Parameters)::Real

Evaluates the score of a design for a given set of parameters.

Parameters

Parameter	Description

design	a Design

params	a parameters object holding the required information about the score

Details

A method score(design::Design)::Real is also implemented which uses the parameter object stored within the design object after optimization. Note that this is only possible if the design was created as result of calling getoptimaldesign().

source

BinaryTwoStageDesigns.test — Method.

test(design::Design, x1::T, x2::T)::Bool where {T<:Integer}

Binary test decision of design when observing x1 responses in stage one and x2 responses in stage two.

Parameters

Parameter	Description

design	a Design

x1	number of stage-one responses

x2	number of stage-two responses

source

BinaryTwoStageDesigns.simulate — Method.

simulate(design::Design, p::T2, nsim::T1) where {T1<:Integer, T2<:Real}

Simulate nsim trials of design design with true response probability p.

Parameters

Parameter	Description

design	a Design

p	true response probability

nsim	number of simulated trials

Return Value

A DataFrame with columns x1 (stage one responses) n (overall sample size) c (critical value), x2 (stage-two responses), and rejectedH0 (test decision).

source

BinaryTwoStageDesigns.jeffreysprior — Method.

jeffreysprior(design::Design)

Computes the Jeffreys prior of any given design.

Parameters

Parameter	Description

design	a Design

Return Value

A function with signature prior{T<:Real}(p::T)::Real where p is the response probability and prior(p) the PDF of the Jeffres prior at p.

Examples

julia> ss = SimpleSampleSpace(10:25, 100, n2min = 5)
julia> params = SimpleMinimalExpectedSampleSize(ss, .2, .4, .05, .2, .4)
julia> design, res = getoptimaldesign(params, solver = Gurobi.GurobiSolver())
julia> f = jeffreysprior(design)
julia> f(.5)

source

Distributions.pdf — Method.

pdf{T1<:Integer, T2<:Real}(design::Design, x1::T1, x2::T1, p::T2)

Probability density function of $(x1, x2)$ responses in stage one and two, respectively, under design given response rate $p$.

Parameters

Parameter	Description

design	a Design

x1	number of stage-one responses

x2	number of stage-two responses

p	response probability

source

BinaryTwoStageDesigns.save — Method.

save(filename::String, design::Design)

Save design as .jls file.

Parameters

Parameter	Description

filename	filename for sving design

design	a Design

source

Base.DataFmt.writecsv — Method.

writecsv(filename::String, design::Design; label::String = "")

Save design as .csv file.

Parameters

Parameter	Description

filename	filename for sving design

design	a Design

label	string name of design

source