Design Parameters

Design Parameters

Overview

Objects of type Parameters are used to define objective functions and their respective parameters for finding optimal two-stage designs. Every Parameters object also contains a SampleSpace which restricts the feasible region of the sample space and implements a score method which is to be optimized.

BinaryTwoStageDesigns.Parameters — Type.

Parameters

Abstract type representing a generic set of paramters for finding optimal two-stage designs.

source

BinaryTwoStageDesigns.label — Method.

label(par::Parameters)

Return the label of a Parameters-object.

source

BinaryTwoStageDesigns.null — Method.

null(par::Parameters)

Return the response rate under the null hypothesis of a Parameters-object.

source

BinaryTwoStageDesigns.mtoer — Method.

mtoer(par::Parameters)

Return the maximal type one error rate of a Parameters-object.

source

BinaryTwoStageDesigns.mcrv — Method.

mcrv(par::Parameters)

Return the minimal clinically relevant response rate of a Parameters-object.

source

BinaryTwoStageDesigns.samplespace — Method.

samplespace(par::Parameters)

Return the SampleSpace-object of a Parameters-object.

source

MESS

BinaryTwoStageDesigns.MESS — Type.

MESS{T_samplespace<:SampleSpace,TR<:Real} <: PointAlternative

MESS{T_samplespace<:SampleSpace}(
  samplespace::T_samplespace,
  p0, p1,
  alpha, beta,
  pess;
  minstoppingforfutility::Real   = 0.0,
  minconditionalpower::Real      = 0.0,
  MONOTONECONDITIONALPOWER::Bool = true
)

This type represents a set of parameters for finding optimal two-stage designs minimizing the expected sample size on a point in the parameter space subject to type one and two error rate constraints.

> [1] Simon R. Optimal two-stage designs for phase II clinical trials. `Controlled Clinical Trials` 1989; 10, 1-10.

> [2] Kunzmann K and Kieser M. Optimal adaptive two-stage designs for single-arm trials with binary endpoint. `arxive.org` 2016; arXiv:1605.00249.

Parameters

Parameter	Description

samplespace	a sample space object

p0	upper boundary of the null hypothesis

p1	point alternative to power on

alpha	maximal tolerable type one error rate

beta	maximal tolerable type two error rate on p1

pess	response rate under which to minimize expected sample size

minstoppingforfutility	minimal probability for stopping for futility under p0

minconditionalpower	minimal conditional power upon continuation to stage two

MONOTONECONDITIONALPOWER	if true, the conditional power must be monotonously increasing, this constraint is only relevant if nmax is set very restrictively

source

LiuScore

BinaryTwoStageDesigns.LiuScore — Type.

LiuScore{TI<:Integer,TR<:Real}

LiuScore(
  samplespace::SampleSpace,
  p0::Real, pmcrv::Real, prior, 
  alpha::Real, beta::Real,
  fs::Real, fp::Real;
  npriorpivots::Integer = 50,
  npivots::Integer = 15,
  minconditionalpower::Real = 0.0,
  MONOTONECONDITIONALPOWER::Bool = false,
  label::String = ""
)

This type represents a set of parameters for finding optimal two-stage designs minimizing a variant of the criterion proposed by Liu et al. in [1].

[1] Liu G, Frank G, Zhu C, and Lu C. Evaluating the adaptive performance of flexible sample size designs with treatment difference in an interval. Statistics in Medicine 2008; 27(4): 584-596.

Parameters

Parameter	Description

samplespace	a sample space object

p0	upper boundary of the null hypothesis

pmcrv	point alternative to power on

prior	prior distribution on true response rate, expected score under the prior conditional on a relevant effect is minimized

alpha	maximal tolerable type one error rate

beta	maximal tolerable type two error rate on p1

fs	scaling parameter as described in [1]

fp	scaling parameter as described in [1]

npriorpivots	number of pivot points for integration of prior

npivots	number of pivot points piecewise linear approximation of local score

minconditionalpower	minimal conditional power upon continuation to stage two

MONOTONECONDITIONALPOWER	if true, the conditional power must be monotonously increasing, this constraint is only relevant if nmax is set very restrictively

label	descriptive string for parameter set

source

EB

BinaryTwoStageDesigns.EB — Type.

EB{TI<:Integer,TR<:Real} <: VagueAlternative

EB( # default values
  samplespace::SampleSpace,
  p0::Real, pmcrv::Real, prior,
  gamma::Real, lambda::Real;
  a::Real = 1, b::Real = 1, targetpower::Real = .8, k::Real = 1,
  alpha::Real = 0.05,
  minconditionalpower::Real = 0.0, MONOTONECONDITIONALPOWER::Bool = true,
  npriorpivots::Integer = 50, ngpivots::Integer = 15,
  label::String = ""
)

This type represents a set of parameters for finding optimal two-stage designs maximizing the expected benefit.

Parameters

Parameter	Description

samplespace	a sample space object

p0	upper boundary of the null hypothesis

pmcrv	point alternative to power on

prior	prior distribution on true response rate, expected score under the prior conditional on a relevant effect is minimized

gamma	per patient costs

lambda	benefit of correctly rejecting null if relevant effect exists

a	scaling parameter of nonlinear power transfrom (set to infinity for thresholding)

b	scaling parameter of nonlinear power transfrom

targetpower	target power for thresholding

k	factor for non-responder costs

alpha	maximal type one error rate

minconditionalpower	minimal conditional power upon continuation to stage two

MONOTONECONDITIONALPOWER	if true, the conditional power must be monotonously increasing, this constraint is only relevant if nmax is set very restrictively

npriorpivots	number of pivot points for integration of prior

npivots	number of pivot points piecewise linear approximation of local score

label	descriptive string for parameter set

source

MBESS

BinaryTwoStageDesigns.MBESS — Type.

MBESS{TI<:Integer,TR<:Real} <: VagueAlternative

MBESS(
  samplespace::SampleSpace,
  p0::Real, pmcrv::Real, prior;
  a::Real = 1, b::Real = 1, targetpower::Real = .8, k::Real = 1,
  alpha::Real = 0.05, beta::Real = 0.2,
  minconditionalpower::Real = 0.0, MONOTONECONDITIONALPOWER::Bool = true,
  npriorpivots::Integer = 50, ngpivots::Integer = 15,
  label::String = ""
)

This type represents a set of parameters for finding optimal two-stage designs minimizing expected sample size subject to a constraint on expected (transformed) power.

Parameters

Parameter	Description

samplespace	a sample space object

p0	upper boundary of the null hypothesis

pmcrv	point alternative to power on

prior	prior distribution on true response rate, expected score under the prior conditional on a relevant effect is minimized

a	scaling parameter of nonlinear power transfrom (set to infinity for thresholding)

b	scaling parameter of nonlinear power transfrom

targetpower	target power for thresholding

k	factor for non-responder costs

alpha	maximal type one error rate

beta	expected transformed power must be larger than 1 - beta

minconditionalpower	minimal conditional power upon continuation to stage two

MONOTONECONDITIONALPOWER	if true, the conditional power must be monotonously increasing, this constraint is only relevant if nmax is set very restrictively

npriorpivots	number of pivot points for integration of prior

npivots	number of pivot points piecewise linear approximation of local score

label	descriptive string for parameter set

source