Design Parameters

Design Parameters

Overview

Objects of type Parameters are used to store parameters required for finding optimal two-stage designs. Every Parameters object contains a sample space which restricts the feasible region of the sample space and implements a score method which is to be optimized.

BinaryTwoStageDesigns.Parameters — Type.

Parameters

Abstract type representing a generic set of paramters for finding optimal two-stage designs.

source

BinaryTwoStageDesigns.PointAlternative — Type.

PointAlternative <: Parameters

Abstract type representing a generic set of paramters for finding optimal two-stage designs which are characterized by a specific point alternative.

source

SimpleMinimalExpectedSampleSize

Currently, only a single parameter type (and thus objective function) is implemented. A SimpleMinimalExpectedSampleSize objects holds information about the null hypothesis and a point alternative on which the power is calculated as well as allowing to specify a response rate under which the expected sample size is to be minimized.

BinaryTwoStageDesigns.SimpleMinimalExpectedSampleSize — Type.

SimpleMinimalExpectedSampleSize{T_samplespace<:SampleSpace} <: PointAlternative

This type represents a set of parameters for finding optimal two-stage designs minimizing the expected sample size on a point in the parameter space subject to type one and two error rate constraints.

source