Sample Spaces

The Sample Space

Sample space objects can be used to encode the feasible space of binary two-stage designs for optimization.

BinaryTwoStageDesigns.SampleSpace — Type.

SampleSpace

Defines restrictions on the sample space of a two-stage design to be used during optimization.

source

BinaryTwoStageDesigns.SampleSpace — Method.

SampleSpace(
  n1range,
  nmax::TI;
  n2min::TI = 1, 
  maxnfact::TR = Inf, 
  nmincont::TI = 0, 
  maxvariables::TI = 500000, 
  GS::Bool = false
) where {TI<:Integer,TR<:Real}

Creates a SampleSpace-object which defines the search space for finding optimal two-stage designs using optimaldesign. The primary use of dedicated sample sapce objects is the incoporation of specific operational constraints on optimal designs (e.g. maximal sample size).

Parameters

Parameter	Description

n1range	possible stage-one sample sizes, can be anything which is convertable to an integer vector

nmax	maximal overall sample size (stage one and two combined)

[n2min]	minimal stage-two sample size

[maxnfact]	maximal overall sample size must be smaller than maxnfact * n1

[nmincont]	minimal sample size upon rejection of the null hypothesis

[maxvariables]	(approximate) maximal number of binary variables used when finding optimal design; if the sample space is too big this can be used to find approximate solutions to the optimization problem

[GS]	flag indicating wheather the design should be group-sequential (constant stage two sample size upon continuation)

source

BinaryTwoStageDesigns.interimsamplesizerange — Method.

interimsamplesizerange(ss::SampleSpace)

Return integer vector of valid $n_1$ values.

source

BinaryTwoStageDesigns.maxsamplesize — Method.

maxsamplesize(ss::SampleSpace)

Return maximal overal sample size as integer.

source

BinaryTwoStageDesigns.maxsamplesize — Method.

maxsamplesize(
  ss::SampleSpace{TI,TR}, n1::TI2
) where {TI<:Integer,TR<:Real,TI2<:Integer}

Return maximal sample size given n1 as integer.

source

BinaryTwoStageDesigns.isgroupsequential — Method.

isgroupsequential(ss::SampleSpace)

Indicate whether the sample space is restricted to group-sequential designs.

source