Sample Spaces

Sample Spaces

Sample spaces are used to encode the feasible space of binary two-stage designs for optimization. Currently, only a single option SimpleSampleSpace exists which allows to specify a range of possible stage-one sample sizes, maximal overall sample size and various nicety constraints.

BinaryTwoStageDesigns.SampleSpace — Type.

SampleSpace

Abstract type representing generic sample spaces defining the search space for finding optimal two stage designs.

source

BinaryTwoStageDesigns.SimpleSampleSpace — Method.

SimpleSampleSpace{T<:Integer}(n1range, nmax::T; n2min::T = 1, maxnfact::Real = Inf, nmincont::T = 0, maxvariables::T = 500000, GS::Bool = false)

Constructs an object of type SimpleSampleSpace defining the search space for finding optimal two stage designs.

Parameters

Parameter	Description

n1range	possible stage-one sample sizes, can be anything which is convertable to an integer vector

n2min	minimal stage-two sample size

nmax	maximal overall sample size (stage one and two combined)

maxnfact	nmax must be smaller than maxnfact*n1

nmincont	minimal sample size upon rejection of the null hypothesis

maxvariables	(approximate) maximal number of binary variables used when finding optimal design; if the sample space is too big this can be used to find approximate solutions to the optimization problem

GS	flag indicating wheather the design should be group-sequential (constant stage two sample size upon continuation)

Return Value

An object of type SimpleSampleSpace with the respective parameters.

Examples

julia> SimpleSampleSpace(10:25, 100, n2min = 5)

source